Propagate value proxy to other widgets
Adding the value proxy to:
 - ctkDoubleSlider
 - ctkSliderWidget
 - ctkCoordinatesWidget
 - ctkRangeWidget
 - ctkDoubleRangeSlider

For the ctkSliderWidget, since it's using the ctkDoubleSlider and the
ctkDoubleSpinBox, we can simply pass the value proxy to these. The same
thing applies to ctkCoordinatesWidget.
For the others, it's like the ctkDoubleSpinBox. The proxy is set between
the values accessors.

The tests added are very similar to the ctkDoubleSpinBox test, just tweaked
to these classes specific API.